Remove unneeded protocols for lookup classes
